Remove `ColumnInterface`
