+++ src/QueryBuilder/QueryBuilderInterface.php
@@ -10,7 +10,6 @@ use Yiisoft\Db\Connection\ServerInfoInterface;
 use Yiisoft\Db\Exception\InvalidArgumentException;
 use Yiisoft\Db\Expression\ExpressionBuilderInterface;
 use Yiisoft\Db\Expression\ExpressionInterface;
-use Yiisoft\Db\Schema\Builder\ColumnInterface;
 use Yiisoft\Db\Schema\Column\ColumnSchemaInterface;
 use Yiisoft\Db\Schema\QuoterInterface;
 
@@ -42,62 +41,12 @@ interface QueryBuilderInterface extends DDLQueryBuilderInterface, DMLQueryBuilde
     /**
      * Builds column definition based on given column instance.
      *
-     * @param ColumnInterface|ColumnSchemaInterface|string $column the column instance or string column definition which
-     * should be converted into a database string representation.
+     * @param ColumnSchemaInterface|string $column the column instance or string column definition which should be
+     * converted into a database string representation.
      *
      * @return string the SQL column definition.
      */
-    public function buildColumnDefinition(ColumnInterface|ColumnSchemaInterface|string $column): string;
-
-    /**
-     * Converts an abstract column type into a physical column type.
-     *
-     * The conversion is done using the type map specified in {@see typeMap}.
-     * The following abstract column types are supported (using MySQL as an example to explain the corresponding
-     * physical types):
-     *
-     * - `pk`: an auto-incremental primary key type, will be converted into "int(11) NOT NULL AUTO_INCREMENT PRIMARY
-     *    KEY"
-     * - `bigpk`: an auto incremental primary key type, will be converted into "bigint(20) NOT NULL AUTO_INCREMENT
-     *    PRIMARY KEY"
-     * - `upk`: an unsigned auto incremental primary key type, will be converted into "int(10) UNSIGNED NOT NULL
-     *    AUTO_INCREMENT PRIMARY KEY"
-     * - `char`: char type, will be converted into "char(1)"
-     * - `string`: string type, will be converted into "varchar(255)"
-     * - `text`: a long string type, will be converted into "text"
-     * - `smallint`: a small integer type, will be converted into "smallint(6)"
-     * - `integer`: integer type, will be converted into "int(11)"
-     * - `bigint`: a big integer type, will be converted into "bigint(20)"
-     * - `boolean`: boolean type, will be converted into "tinyint(1)"
-     * - `float``: float number type, will be converted into "float"
-     * - `decimal`: decimal number type, will be converted into "decimal"
-     * - `datetime`: datetime type, will be converted into "datetime"
-     * - `timestamp`: timestamp type, will be converted into "timestamp"
-     * - `time`: time type, will be converted into "time"
-     * - `date`: date type, will be converted into "date"
-     * - `money`: money type, will be converted into "decimal(19,4)"
-     * - `binary`: binary data type, will be converted into "blob"
-     *
-     * If the abstract type has two or more parts separated by spaces (such as "string NOT NULL"), then only the first
-     * part will be converted, and the rest of the parts will be appended to the converted result.
-     *
-     * For example, 'string NOT NULL' is converted to 'varchar(255) NOT NULL'.
-     *
-     * For some abstract types, you can also specify a length or precision constraint by appending it in round brackets
-     * directly to the type.
-     *
-     * For example, `string(32)` will be converted into "varchar(32)" on a MySQL database.
-     * If the underlying DBMS doesn't support these kinds of constraints for a type, it will be ignored.
-     *
-     * If a type can't be found in {@see typeMap}, it will be returned without any change.
-     *
-     * @param ColumnInterface|string $type Abstract column type.
-     *
-     * @return string Physical column type.
-     *
-     * @deprecated Use {@see buildColumnDefinition()}. Will be removed in version 2.0.
-     */
-    public function getColumnType(ColumnInterface|string $type): string;
+    public function buildColumnDefinition(ColumnSchemaInterface|string $column): string;
 
     /**
      * Returns the column definition builder for the current DBMS.
